backend/pyproject.toml
[project]
name = "story-reels"
version = "0.1.0"
description = "Automated Story-to-Reels App for Kids"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "sqlalchemy>=2.0.28",
    "alembic>=1.13.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "moviepy>=1.0.3",
    "pillow>=10.2.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "aiofiles>=23.2.1",
    "websockets>=12.0",
    "openai>=1.14.0",
    "numpy>=1.26.0",
    "python-bidi>=0.4.2",
    "arabic-reshaper>=3.0.0",
    "psycopg2-binary>=2.9.9",
    "pyarabic>=0.6.15",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "black>=24.2.0",
    "ruff>=0.3.0",
]

[tool.setuptools]
packages = ["app"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
